feat: Implement TransactionGreater rules throught makeRuleComponent
